Cleanup in plugin_unloaded and on_exit
You can register an exit handler in the `acquire` function. These exit handlers
are kept in a dict to be used in the `release` function. Users must call
`release` in `plugin_unloaded` and `on_exit` of an event listener.
The loop is stopped in `release`. If we're dealing with app shutdown then we
invoke all exit handlers. Otherwise we invoke the exit handler of the plugin
that called `release`, if any. Such exit handlers are always ran from the main
thread. This ensures proper cleanup of unloading packages.